change tasks to rj_smtr
