+++ pipelines/rj_smtr/veiculo/flows.py
@@ -18,8 +18,6 @@ from pipelines.utils.tasks import (
     rename_current_flow_run_now_time,
     get_current_flow_mode,
     get_current_flow_labels,
-    get_yesterday,
-    get_previous_date,
 )
 
 # SMTR Imports #
@@ -42,6 +40,8 @@ from pipelines.rj_smtr.tasks import (
     fetch_dataset_sha,
     get_run_dates,
     get_join_dict,
+    get_yesterday,
+    get_previous_date,
 )
 
 from pipelines.rj_smtr.veiculo.tasks import (
